pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "matplotlib>=3.10.1",
    "newsapi-python>=0.2.7",
    "numpy>=2.2.3",
    "pandas>=2.2.3",
    "plotly>=6.0.0",
    "scikit-learn>=1.6.1",
    "streamlit>=1.42.2",
    "textblob>=0.19.0",
    "yfinance>=0.2.54",
]
